+++ CHANGELOG.md
@@ -3,9 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
-- fix: Add '.js' file extensions to local ESM import paths
 
-## 0.1.0 – 0.1.1
+## 0.1.0 – 0.1.2
 
 _2023-03-24_